add support for macOS less-than-10.12

+++ di-hazel.sh
@@ -4,9 +4,7 @@
 # From:	Tj Luo.ma
 # Mail:	luomat at gmail dot com
 # Web: 	http://RhymesWithDiploma.com
-# Date:	2015-10-23
-
-## 2016-04-22 - changed from .dmg to .zip
+# Date:	2015-10-23 (2016-04-22 - changed from .dmg to .zip)
 
 NAME="$0:t:r"
 
@@ -93,32 +91,60 @@ function check_install_location {
 	fi
 }
 
-	# If there's no installed version, output 4.0.0 so the Sparkle feed will give us the proper download URL
-	## DO NOT SET TO ZERO
-INSTALLED_VERSION=`defaults read ${INSTALL_TO}/Contents/Info CFBundleShortVersionString 2>/dev/null || echo '4.0.0'`
+OS_VER=$(sw_vers -productVersion | cut -d. -f2)
 
-XML_FEED="https://www.noodlesoft.com/Products/Hazel/generate-appcast.php?version=$INSTALLED_VERSION"
-
-INFO=($(curl -sfL "$XML_FEED" \
-			| tr -s ' ' '\012' \
-			| egrep '^(sparkle:version|url)=' \
-			| head -2 \
-			| awk -F'"' '/=/{print $2}'))
-
-LATEST_VERSION="$INFO[1]"
-
-URL="$INFO[2]"
-
-	# If any of these are blank, we should not continue
-if [ "$INFO" = "" -o "$LATEST_VERSION" = "" -o "$URL" = "" ]
+if [[ "$OS_VER" -lt "12" ]]
 then
-	echo "$NAME: Error: bad data received:
-	INFO: $INFO
-	LATEST_VERSION: $LATEST_VERSION
-	URL: $URL
-	"
 
-	exit 1
+	## the 'official' URL for version 4.3.5 is a DMG:
+	# 	URL='https://www.noodlesoft.com/Downloads/Hazel-4.3.5.dmg'
+	## but this script assumes a ZIP file not a DMG. My personal archived version is at
+	# 	URL='https://iusethis.luo.ma/hazel/old/Hazel-4.3.5.zip'
+	## but the one from Hazel's Sparkle feed is currently still available at
+	# 	URL='https://s3.amazonaws.com/Noodlesoft/Hazel-sparkle-4.3.5.zip'
+	## so we'll use that one for now.
+	#
+	## Checksum for Hazel-4.3.5.zip / Hazel-sparkle-4.3.5.zip:
+	#
+	# sha256 = da9243ee965dbbbed1483373c047b48974135889db037e94fee15e2e695d1d9a
+	#    md5 = f77fd65a61d6b4164600f1e7e516b946
+
+	URL='https://s3.amazonaws.com/Noodlesoft/Hazel-sparkle-4.3.5.zip'
+	LATEST_VERSION='4.3.5'
+	RELEASE_NOTES='no'
+
+else
+
+	RELEASE_NOTES='yes'
+
+		# If there's no installed version, output 4.0.0 so the Sparkle feed will give us the proper download URL
+		## DO NOT SET TO ZERO
+	INSTALLED_VERSION=`defaults read ${INSTALL_TO}/Contents/Info CFBundleShortVersionString 2>/dev/null || echo '4.0.0'`
+
+	XML_FEED="https://www.noodlesoft.com/Products/Hazel/generate-appcast.php?version=$INSTALLED_VERSION"
+
+	INFO=($(curl -sfL "$XML_FEED" \
+				| tr -s ' ' '\012' \
+				| egrep '^(sparkle:version|url)=' \
+				| head -2 \
+				| awk -F'"' '/=/{print $2}'))
+
+	LATEST_VERSION="$INFO[1]"
+
+	URL="$INFO[2]"
+
+		# If any of these are blank, we should not continue
+	if [ "$INFO" = "" -o "$LATEST_VERSION" = "" -o "$URL" = "" ]
+	then
+		echo "$NAME: Error: bad data received:
+		INFO: $INFO
+		LATEST_VERSION: $LATEST_VERSION
+		URL: $URL
+		"
+
+		exit 1
+	fi
+
 fi
 
 if [[ -e "$INSTALL_TO" ]]
@@ -146,40 +172,96 @@ fi
 
 FILENAME="$HOME/Downloads/$INSTALL_TO:t:r-$LATEST_VERSION.zip"
 
-if (( $+commands[lynx] ))
+if [[ "$RELEASE_NOTES" == "yes" ]]
 then
+	if (( $+commands[lynx] ))
+	then
 
-	RELEASE_NOTES_URL='https://www.noodlesoft.com/release_notes'
+		RELEASE_NOTES_URL='https://www.noodlesoft.com/release_notes'
 
-	( echo -n "$NAME: Release Notes for Hazel " ;
-		(curl -sfL "$RELEASE_NOTES_URL" | sed '1,/<h1>/d; /<\/ul>/,$d' ; echo '</ul>') |\
-		lynx -dump -nomargins -width=10000 -assume_charset=UTF-8 -pseudo_inlines -stdin ;
-		echo "\nSource: <${RELEASE_NOTES_URL}>\nURL: $URL" ) | tee "$FILENAME:r.txt"
+		( echo -n "$NAME: Release Notes for Hazel " ;
+			(curl -sfL "$RELEASE_NOTES_URL" | sed '1,/<h1>/d; /<\/ul>/,$d' ; echo '</ul>') |\
+			lynx -dump -nomargins -width=10000 -assume_charset=UTF-8 -pseudo_inlines -stdin ;
+			echo "\nSource: <${RELEASE_NOTES_URL}>\nURL: $URL" ) | tee "$FILENAME:r.txt"
+	fi
 fi
 
-	# Server does not support continued downloads, so assume that this is incomplete and try again
-[[ -f "$FILENAME" ]] && rm -f "$FILENAME"
 
-echo "$NAME: Downloading '$URL' to '$FILENAME':"
+if [[ -f "$FILENAME" ]]
+then
 
-curl --continue-at - --fail --location --output "$FILENAME" "$URL"
+	## The server does not support resuming downloads, so if we have the zip file, we need to test it
+	## to see if it is valid. If yes, we keep it. If no, we delete it and download fresh
 
-EXIT="$?"
+	(command unzip -l "$FILENAME" 2>&1 )>/dev/null
 
-	## exit 22 means 'the file was already fully downloaded'
-[ "$EXIT" != "0" -a "$EXIT" != "22" ] && echo "$NAME: Download of $URL failed (EXIT = $EXIT)" && exit 0
+	EXIT="$?"
 
-[[ ! -e "$FILENAME" ]] && echo "$NAME: $FILENAME does not exist." && exit 0
+	if [ "$EXIT" = "0" ]
+	then
+		echo "$NAME: '$FILENAME' is a valid zip file."
 
-[[ ! -s "$FILENAME" ]] && echo "$NAME: $FILENAME is zero bytes." && rm -f "$FILENAME" && exit 0
+	else
+		echo "$NAME: '$FILENAME' is an invalid zip file (\$EXIT = $EXIT)"
 
-(cd "$FILENAME:h" ; echo "\nLocal sha256:" ; shasum -a 256 -p "$FILENAME:t" ) >>| "$FILENAME:r.txt"
+		mv -fv "$FILENAME" "$HOME/.Trash/"
 
-# If we get here we are ready to install
-# let's make sure we can
+		mv -fv "$FILENAME:r".* "$HOME/.Trash/"
 
+	fi
+
+fi
+
+## so now if we do NOT have the filename, then we try to download it
+
+if [[ ! -f "$FILENAME" ]]
+then
+
+		# Server does not support continued downloads, so assume that this is incomplete and try again
+	[[ -f "$FILENAME" ]] && rm -f "$FILENAME"
+
+	echo "$NAME: Downloading '$URL' to '$FILENAME':"
+
+	curl --continue-at - --fail --location --output "$FILENAME" "$URL"
+
+	EXIT="$?"
+
+		## exit 22 means 'the file was already fully downloaded'
+	[ "$EXIT" != "0" -a "$EXIT" != "22" ] && echo "$NAME: Download of $URL failed (EXIT = $EXIT)" && exit 0
+
+	[[ ! -e "$FILENAME" ]] && echo "$NAME: $FILENAME does not exist." && exit 0
+
+	[[ ! -s "$FILENAME" ]] && echo "$NAME: $FILENAME is zero bytes." && rm -f "$FILENAME" && exit 0
+
+	(cd "$FILENAME:h" ; echo "\nLocal sha256:" ; shasum -a 256 -p "$FILENAME:t" ) >>| "$FILENAME:r.txt"
+
+		## make sure that the .zip is valid before we proceed
+	(command unzip -l "$FILENAME" 2>&1 )>/dev/null
+
+	EXIT="$?"
+
+	if [ "$EXIT" = "0" ]
+	then
+		echo "$NAME: '$FILENAME' is a valid zip file."
+
+	else
+		echo "$NAME: '$FILENAME' is an invalid zip file (\$EXIT = $EXIT)"
+
+		mv -fv "$FILENAME" "$HOME/.Trash/"
+
+		mv -fv "$FILENAME:r".* "$HOME/.Trash/"
+
+		exit 0
+
+	fi
+
+fi
+
+	# If we get here we are ready to install
+	# let's make sure we can
 check_install_location
 
+	## unzip to a temporary directory
 UNZIP_TO=$(mktemp -d "${TMPDIR-/tmp/}${NAME}-XXXXXXXX")
 
 echo "$NAME: Unzipping '$FILENAME' to '$UNZIP_TO':"
@@ -216,9 +298,14 @@ fi
 
 if [[ -e "$INSTALL_TO" ]]
 then
-	echo "$NAME: Moving existing (old) \"$INSTALL_TO\" to \"$HOME/.Trash/\"."
 
-	mv -vf "$INSTALL_TO" "$HOME/.Trash/$INSTALL_TO:t:r.$INSTALLED_VERSION.prefPane"
+	pgrep -xq "$INSTALL_TO:t:r" \
+	&& LAUNCH='yes' \
+	&& osascript -e "tell application \"$INSTALL_TO:t:r\" to quit"
+
+	echo "$NAME: Moving existing (old) '$INSTALL_TO' to '$HOME/.Trash/'."
+
+	mv -vf "$INSTALL_TO" "$HOME/.Trash/$INSTALL_TO:t:r.$INSTALLED_VERSION.app"
 
 	EXIT="$?"
